fix: Fix build error - Move Suspense boundary to parent component
